requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
fabric==2.4.0
idna==2.8
invoke==1.2.0
paramiko==2.6.0
pycparser==2.19
PyNaCl==1.3.0
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
urllib3==1.25.3
